+++ cli/cmd/completion.go
@@ -1,12 +1,14 @@
 package cmd
 
 import (
+	"bytes"
 	"fmt"
-	"os"
 
 	"github.com/spf13/cobra"
 	"github.com/tarantool/tt/cli/cmdcontext"
 	"github.com/tarantool/tt/cli/modules"
+	"github.com/tarantool/tt/cli/rocks"
+	"github.com/tarantool/tt/cli/util"
 )
 
 // NewCompletionCmd creates a new completion command.
@@ -46,17 +48,54 @@ func RootShellCompletionCommands(cmd *cobra.Command, args []string,
 	return commands, cobra.ShellCompDirectiveDefault
 }
 
+// injectRocksCompletion combines luarocks completions with cobra completions.
+func injectRocksCompletion(shell string, completion []byte) ([]byte, error) {
+	label := []byte(`    # The user could have moved the cursor backwards on the command-line.`)
+
+	injection, err := util.ReadEmbedFile(rocks.EmbedCompletions, "completions/"+shell+"_injection")
+	if err != nil {
+		return nil, err
+	}
+	rocks, err := util.ReadEmbedFile(rocks.EmbedCompletions, "completions/"+shell+"_rocks")
+	if err != nil {
+		return nil, err
+	}
+
+	res := bytes.Buffer{}
+	idx := bytes.Index(completion, label)
+	if idx == -1 {
+		return nil, fmt.Errorf("failed to inject LuaRocks completions")
+	}
+	res.Write(completion[:idx])
+	res.Write([]byte(injection))
+	res.Write(completion[idx:])
+	res.Write([]byte(rocks))
+
+	return res.Bytes(), nil
+}
+
 // internalCompletionCmd is a default (internal) completion module function.
 func internalCompletionCmd(cmdCtx *cmdcontext.CmdCtx, args []string) error {
+	var buf bytes.Buffer
 	switch shell := args[0]; shell {
 	case "bash":
-		if err := rootCmd.GenBashCompletionV2(os.Stdout, true); err != nil {
+		if err := rootCmd.GenBashCompletionV2(&buf, true); err != nil {
 			return err
 		}
+		res, err := injectRocksCompletion(shell, buf.Bytes())
+		if err != nil {
+			return err
+		}
+		fmt.Print(string(res))
 	case "zsh":
-		if err := rootCmd.GenZshCompletion(os.Stdout); err != nil {
+		if err := rootCmd.GenZshCompletion(&buf); err != nil {
 			return err
 		}
+		res, err := injectRocksCompletion(shell, buf.Bytes())
+		if err != nil {
+			return err
+		}
+		fmt.Print(string(res))
 	default:
 		return fmt.Errorf("specified shell type is not is not supported. Available: bash | zsh")
 	}